requirements.txt
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
dbus-python==1.3.2
filelock==3.13.1
fsspec==2023.12.1
idna==3.6
Jinja2==3.1.2
llvmlite==0.41.1
MarkupSafe==2.1.3
more-itertools==10.1.0
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.2
openai-whisper==20231117
PyAudio==0.2.14
pycairo==1.25.1
pycparser==2.21
PyGObject==3.46.0
regex==2023.10.3
requests==2.31.0
soundfile==0.12.1
SpeechRecognition==3.10.1
sympy==1.12
tiktoken==0.5.2
torch==2.1.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
